Refactor package name from "exntensions" to "extensions"
