+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/extensions/RandomExtensions.kt
@@ -1,4 +1,4 @@
-package dev.teogor.sudoklify.exntensions
+package dev.teogor.sudoklify.extensions
 
 import kotlin.random.Random
