+++ sudoklify/src/main/kotlin/dev/teogor/sudoklify/extensions/SudokuBoardExtensions.kt
@@ -1,4 +1,4 @@
-package dev.teogor.sudoklify.exntensions
+package dev.teogor.sudoklify.extensions
 
 import dev.teogor.sudoklify.types.Board
 import dev.teogor.sudoklify.types.SudokuString
